# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "subjex",
    "description": "JSON API behind the operator pages — 操作页背后的 JSON 接口",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "operatorBearer": []
    },
    {
      "operatorBasic": []
    }
  ],
  "paths": {
...
    "/api/v1/forms/{formKey}/submissions": {
      "get": {
        "tags": [
          "form-submission-endpoint"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "formKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FormSubmissionListDocument"
                }
              }
            }
          },
          "401": {
            "description": "No operator signed in — 没有已登录的操作员"
          },
          "403": {
            "description": "Operator lacks the permission — 操作员缺少这项权限"
          }
        }
      },
      "post": {
        "tags": [
          "form-submission-endpoint"
        ],
        "operationId": "submit",
        "parameters": [
          {
            "name": "formKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormSubmissionDocument"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FormSubmissionResultDocument"
                }
              }
            }
          },
          "401": {
            "description": "No operator signed in — 没有已登录的操作员"
          },
          "403": {
            "description": "Operator lacks the permission — 操作员缺少这项权限"
          }
        }
      }
    },
...
    "/api/v1/forms": {
      "get": {
        "tags": [
          "forms-api-endpoint"
        ],
        "operationId": "index_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FormsIndexDocument"
                }
              }
            }
          },
          "401": {
            "description": "No operator signed in — 没有已登录的操作员"
          },
          "403": {
            "description": "Operator lacks the permission — 操作员缺少这项权限"
          }
        }
      }
    },
    "/api/v1/forms/{formKey}": {
      "get": {
        "tags": [
          "forms-api-endpoint"
        ],
        "operationId": "detail_1",
        "parameters": [
          {
            "name": "formKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FormsDocument"
                }
              }
            }
          },
          "401": {
            "description": "No operator signed in — 没有已登录的操作员"
          },
          "403": {
            "description": "Operator lacks the permission — 操作员缺少这项权限"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FormSubmissionDocument": {
        "type": "object",
        "properties": {
          "values": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "EffectOutcomeDocument": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "outcome": {
            "type": "string"
          }
        }
      },
      "FormSubmissionResultDocument": {
        "type": "object",
        "properties": {
          "submissionId": {
            "type": "string"
          },
          "formKey": {
            "type": "string"
          },
          "declarationVersion": {
            "type": "integer",
            "format": "int32"
          },
          "resultSummary": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "effects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EffectOutcomeDocument"
            }
          }
        }
      },
...
      "FormIndexDocument": {
        "type": "object",
        "properties": {
          "formKey": {
            "type": "string"
          },
          "titleZh": {
            "type": "string"
          },
          "titleEn": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "permission": {
            "type": "string"
          },
          "tenantScoped": {
            "type": "boolean"
          },
          "domainAction": {
            "type": "string"
          },
          "entityKey": {
            "type": "string",
            "nullable": true,
            "description": "Optional entity key when domainAction is entity.record.upsert"
          }
        }
      },
      "FormsIndexDocument": {
        "type": "object",
        "properties": {
          "forms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FormIndexDocument"
            }
          }
        }
      },
      "FieldDocument": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "enumValues": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Allowed values when kind is enum; empty otherwise"
          }
        }
      },
      "FormsDocument": {
        "type": "object",
        "properties": {
          "formKey": {
            "type": "string"
          },
          "titleZh": {
            "type": "string"
          },
          "titleEn": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "permission": {
            "type": "string"
          },
          "tenantScoped": {
            "type": "boolean"
          },
          "domainAction": {
            "type": "string"
          },
          "entityKey": {
            "type": "string",
            "nullable": true,
            "description": "Optional entity key when domainAction is entity.record.upsert"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldDocument"
            }
          }
        }
      },
      "FormSubmissionHistoryDocument": {
        "type": "object",
        "properties": {
          "submissionId": {
            "type": "string"
          },
          "formKey": {
            "type": "string"
          },
          "declarationVersion": {
            "type": "integer",
            "format": "int32"
          },
          "actorIdentityId": {
            "type": "string"
          },
          "loginName": {
            "type": "string"
          },
          "valuesJson": {
            "type": "string"
          },
          "resultSummary": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FormSubmissionListDocument": {
        "type": "object",
        "properties": {
          "submissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FormSubmissionHistoryDocument"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "operatorBasic": {
        "type": "http",
        "description": "Operator login and password (scripts/local tooling) — 操作员口令（脚本/本机工具）",
        "scheme": "basic"
      },
      "operatorBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Opaque access token from POST /api/v1/auth/login — 来自登录接口的不透明访问令牌"
      }
    }
  }
}